fix: fix cabinet files structure and name of components refactoring
