+++ components/cabinet/graphs/repositories.ts
@@ -19,21 +19,3 @@ export const REPOS = gql`
     }
   }
 `;
-
-export const GET_ALL_BRANCHES_IN_REPO = gql`
-  query FindBrunches($repo_name: String!, $login: String!) {
-    repository(owner: $login, name: $repo_name) {
-      refs(
-        refPrefix: "refs/heads/"
-        orderBy: { direction: DESC, field: TAG_COMMIT_DATE }
-        first: 20
-      ) {
-        edges {
-          node {
-            name
-          }
-        }
-      }
-    }
-  }
-`;
